Moved the map logic to viewmodel at EventDetails fragment
