+++ app/src/main/java/com/prokarma/reference/architecture/feature/details/DetailsViewModel.java
@@ -8,9 +8,16 @@ import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
+import com.google.android.gms.maps.CameraUpdate;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
 import com.prokarma.reference.architecture.feature.search.event_list.EventActions;
 import com.prokarma.reference.architecture.model.Embedded;
 import com.prokarma.reference.architecture.model.Event;
+import com.prokarma.reference.architecture.model.Location;
 import com.prokarma.reference.architecture.model.SearchEventsResponse;
 import com.prokarma.reference.architecture.networking.ApplicationDataRepository;
 import com.prokarma.reference.architecture.networking.OnCallListener;
@@ -22,8 +29,11 @@ import java.util.List;
 /**
  * View model in charge of {@link Event} handling.
  */
-public class DetailsViewModel extends ViewModel {
+public class DetailsViewModel extends ViewModel implements OnMapReadyCallback {
     private MutableLiveData<Event> mEventLiveData;
+
+    private GoogleMap mMap;
+    private float DEFAULT_ZOOM = 10;
 
     public DetailsViewModel() {
 
@@ -31,6 +41,7 @@ public class DetailsViewModel extends ViewModel {
 
     public void setEventData(Event mEvent){
         mEventLiveData.setValue(mEvent);
+        updateEventLocOnMap();
     }
 
     public MutableLiveData<Event> getEvent() {
@@ -40,5 +51,26 @@ public class DetailsViewModel extends ViewModel {
             mEventLiveData = new MutableLiveData<>();
         }
         return mEventLiveData;
+    }
+
+    @Override
+    public void onMapReady(GoogleMap googleMap) {
+        mMap = googleMap;
+        updateEventLocOnMap();
+    }
+
+    public void updateEventLocOnMap() {
+        Event event = getEvent().getValue();
+        if(event.embedded.venues.size() > 0 && mMap != null){
+            //Updates event location when map is ready.
+            Location eventLocation = event.embedded.venues.get(0).location;
+            LatLng eventLatLng = new LatLng(Double.valueOf(eventLocation.latitude),
+                    Double.valueOf(eventLocation.longitude));
+            CameraUpdate moveCamera = CameraUpdateFactory.newLatLngZoom(eventLatLng,
+                    DEFAULT_ZOOM);
+            mMap.moveCamera(moveCamera);
+
+            mMap.addMarker(new MarkerOptions().position(eventLatLng));
+        }
     }
 }
