+++ app/src/main/java/com/prokarma/reference/architecture/feature/details/EventDetailsFragment.java
@@ -25,10 +25,9 @@ import com.prokarma.reference.architecture.feature.search.list.ListViewModel;
 import com.prokarma.reference.architecture.model.Event;
 import com.prokarma.reference.architecture.model.Location;
 
-public class EventDetailsFragment extends Fragment implements OnMapReadyCallback{
+public class EventDetailsFragment extends Fragment {
     private FragmentDetailsBinding mBinding;
     private DetailsViewModel mDetailsViewModel;
-    private float DEFAULT_ZOOM = 10;
     private GoogleMap mMap = null;
 
     @Override
@@ -47,7 +46,6 @@ public class EventDetailsFragment extends Fragment implements OnMapReadyCallback
 
         mDetailsViewModel.getEvent().observe(this, (Event event) -> {
             mBinding.setEvent(event);
-            updateEventLocOnMap(event);
         });
 
         Event event = getArguments() != null ? (Event)getArguments().getSerializable("event") : null;
@@ -60,30 +58,10 @@ public class EventDetailsFragment extends Fragment implements OnMapReadyCallback
         super.onActivityCreated(savedInstanceState);
 
         SupportMapFragment mapFragment = (SupportMapFragment) getChildFragmentManager().findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-    }
-
-    private void updateEventLocOnMap(Event event) {
-        if(event.embedded.venues.size() > 0 && mMap != null){
-            //Updates event location when map is ready.
-            Location eventLocation = event.embedded.venues.get(0).location;
-            LatLng eventLatLng = new LatLng(Double.valueOf(eventLocation.latitude),
-                    Double.valueOf(eventLocation.longitude));
-            CameraUpdate moveCamera = CameraUpdateFactory.newLatLngZoom(eventLatLng,
-                    DEFAULT_ZOOM);
-            mMap.moveCamera(moveCamera);
-
-            mMap.addMarker(new MarkerOptions().position(eventLatLng));
-        }
+        mapFragment.getMapAsync(mDetailsViewModel);
     }
 
     DetailsViewModel createViewModel() {
         return ViewModelProviders.of(this).get(DetailsViewModel.class);
-    }
-
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-        mMap = googleMap;
-        updateEventLocOnMap(mDetailsViewModel.getEvent().getValue());
     }
 }
